requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cli-helpers==2.3.0
click==8.1.3
configobj==5.0.8
contourpy==1.0.7
cryptography==36.0.2
cycler==0.11.0
fonttools==4.39.2
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mycli==1.26.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
paramiko==3.1.0
Pillow==9.4.0
prettytable==3.6.0
prompt-toolkit==3.0.38
psutil==5.9.4
pyaes==1.6.1
pycparser==2.21
pyecharts==2.0.2
Pygments==2.14.0
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.2
simplejson==3.18.4
six==1.16.0
speedtest-cli==2.1.3
sqlglot==11.4.3
sqlparse==0.4.3
tabulate==0.9.0
wcwidth==0.2.6
zipp==3.15.0
